wishlist extension working
