+++ react/example-customer-account--wishlist--react/extensions/customer-account-wishlist-extension/.gitignore
@@ -1,0 +1,1 @@
+/dist
