+++ react/example-customer-account--wishlist--react/extensions/customer-account-wishlist-extension/README.md
@@ -1,0 +1,3 @@
+# Customer account UI Extension
+
+Allow developers to build extensions on customer account.
