+++ react/example-customer-account--wishlist--react/extensions/customer-account-wishlist-extension/locales/en.default.json
@@ -1,0 +1,8 @@
+{
+  "wishlist": {
+    "heading": "Wishlist",
+    "addToCart": "Add to cart",
+    "remove": "Remove",
+    "seeWishlistButton": "See wishlist"
+  }
+}
