+++ react/example-customer-account--wishlist--react/extensions/customer-account-wishlist-extension/src/ProfileBlockExtension.tsx
@@ -1,0 +1,16 @@
+import { Banner, reactExtension, useTranslate, useApi, Button } from '@shopify/ui-extensions-react/customer-account';
+
+  export default reactExtension(
+    'customer-account.profile.block.render',
+     () => <BlockExtension />
+  );
+
+  function BlockExtension() {
+    const translate = useTranslate();
+
+    return (
+      <Button to="/">
+        {translate('wishlist.seeWishlistButton')}
+      </Button>
+    );
+  }
